extracted adapter and refactored code
