+++ app/src/main/java/com/joanfuentes/cleanhero/presentation/presenter/ComicListPresenter.java
@@ -1,6 +1,6 @@
 package com.joanfuentes.cleanhero.presentation.presenter;
 
-import com.joanfuentes.cleanhero.presentation.view.ItemListActivity;
+import com.joanfuentes.cleanhero.presentation.view.ComicListActivity;
 import com.joanfuentes.cleanhero.domain.model.Comic;
 import com.joanfuentes.cleanhero.domain.usecase.GetComicsUseCase;
 
@@ -8,12 +8,12 @@ import java.util.List;
 
 import javax.inject.Inject;
 
-public class DashboardPresenter extends BasePresenter {
-    private final ItemListActivity activity;
+public class ComicListPresenter extends BasePresenter {
+    private final ComicListActivity activity;
     private final GetComicsUseCase usecase;
 
     @Inject
-    public DashboardPresenter(ItemListActivity activity, GetComicsUseCase usecase) {
+    public ComicListPresenter(ComicListActivity activity, GetComicsUseCase usecase) {
         this.activity = activity;
         this.usecase = usecase;
     }
