+++ app/src/main/java/com/joanfuentes/cleanhero/presentation/view/internal/di/ActivityComponent.java
@@ -1,9 +1,9 @@
 package com.joanfuentes.cleanhero.presentation.view.internal.di;
 
-import com.joanfuentes.cleanhero.presentation.view.ItemDetailFragment;
-import com.joanfuentes.cleanhero.presentation.view.ItemListActivity;
+import com.joanfuentes.cleanhero.presentation.view.ComicDetailFragment;
+import com.joanfuentes.cleanhero.presentation.view.ComicListActivity;
 
 public interface ActivityComponent {
-        void inject(ItemListActivity activity);
-        void inject(ItemDetailFragment itemDetailFragment);
+        void inject(ComicListActivity activity);
+        void inject(ComicDetailFragment comicDetailFragment);
 }
