+++ app/src/main/java/com/joanfuentes/cleanhero/presentation/view/internal/di/ActivityModule.java
@@ -1,7 +1,7 @@
 package com.joanfuentes.cleanhero.presentation.view.internal.di;
 
-import com.joanfuentes.cleanhero.presentation.view.ItemListActivity;
+import com.joanfuentes.cleanhero.presentation.view.ComicListActivity;
 
 interface ActivityModule {
-    ItemListActivity provideItemListActivity();
+    ComicListActivity provideItemListActivity();
 }
