+++ app/src/main/java/com/joanfuentes/cleanhero/presentation/view/internal/di/RuntimeActivityModule.java
@@ -2,8 +2,8 @@ package com.joanfuentes.cleanhero.presentation.view.internal.di;
 
 import android.app.Activity;
 
-import com.joanfuentes.cleanhero.presentation.view.ItemDetailFragment;
-import com.joanfuentes.cleanhero.presentation.view.ItemListActivity;
+import com.joanfuentes.cleanhero.presentation.view.ComicDetailFragment;
+import com.joanfuentes.cleanhero.presentation.view.ComicListActivity;
 import com.joanfuentes.cleanhero.data.api.endpoint.internal.di.EndpointsModule;
 
 import dagger.Module;
@@ -12,18 +12,18 @@ import dagger.Provides;
 @Module(includes = {EndpointsModule.class})
 public class RuntimeActivityModule implements ActivityModule {
     private Activity activity;
-    private ItemDetailFragment itemDetailFragment;
+    private ComicDetailFragment comicDetailFragment;
 
     public RuntimeActivityModule(Activity activity) {
         this.activity = activity;
     }
 
-    public RuntimeActivityModule(ItemDetailFragment itemDetailFragment) {
-        this.itemDetailFragment = itemDetailFragment;
+    public RuntimeActivityModule(ComicDetailFragment comicDetailFragment) {
+        this.comicDetailFragment = comicDetailFragment;
     }
 
     @Override @Provides
-    public ItemListActivity provideItemListActivity() {
-        return (ItemListActivity) activity;
+    public ComicListActivity provideItemListActivity() {
+        return (ComicListActivity) activity;
     }
 }
