+++ app/src/main/res/layout-w900dp/item_list.xml
@@ -8,7 +8,7 @@
     android:divider="?android:attr/dividerHorizontal"
     android:orientation="horizontal"
     android:showDividers="middle"
-    tools:context="com.joanfuentes.cleanhero.presentation.view.ItemListActivity">
+    tools:context="com.joanfuentes.cleanhero.presentation.view.ComicListActivity">
 
     <!--
     This layout is a two-pane layout for the Items
@@ -25,7 +25,7 @@
         android:layout_marginRight="16dp"
         app:layoutManager="GridLayoutManager"
         app:spanCount="@integer/gridview_columns_number"
-        tools:context="com.joanfuentes.cleanhero.presentation.view.ItemListActivity"
+        tools:context="com.joanfuentes.cleanhero.presentation.view.ComicListActivity"
         tools:listitem="@layout/item_list_content" />
 
     <FrameLayout
