+++ app/src/main/res/layout/activity_item_detail.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.joanfuentes.cleanhero.presentation.view.ItemDetailActivity"
+    tools:context="com.joanfuentes.cleanhero.presentation.view.ComicDetailActivity"
     tools:ignore="MergeRootFrame">
 
     <android.support.design.widget.AppBarLayout
